Se quitan comentarios de TestPromotion. Se corrige el codigo de TestCasting para que compile. Se crean empleados en testEmployee. Se crean constructores en Department, Country, Job, Location. En Employee se valido el sueldo que se envia con respecto al sueldo mayo, sueldo menor

+++ src/modulo4/cast/TestPromotion.java
@@ -8,15 +8,15 @@ package modulo4.cast;
  *
  * @author Humberto
 /*
- *//*
+ */
 
 public class TestPromotion {
 
     public static void main(String[] args) {
         long bigval = 6;    // 6 is an int type, OK
-        int smallval = 99L; // 99L is a long, illegal
+        int smallval = (int)99L; // 99L is a long, illegal
         double z = 12.414F; // 12.414F is float, OK
-        float z1 = 12.414;  // 12.414 is double, illegal
+        float z1 = (float)12.414;  // 12.414 is double, illegal
     }
 }
-*/
+
